// Frontend/Angular/src/app/clients/teacher/model/assignmentStatus.ts
/**
 * Teacher api
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: v1
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */

export type AssignmentStatus = 0 | 1 | 2 | 3;

export const AssignmentStatus = {
    NUMBER_0: 0 as AssignmentStatus,
    NUMBER_1: 1 as AssignmentStatus,
    NUMBER_2: 2 as AssignmentStatus,
    NUMBER_3: 3 as AssignmentStatus
};